`.every` function full implementation, not final
